Initial version of the following functions:
- Ics_RaOpen
- Ics_RaReadOrWrite
- Ics_RaClose
- Ics_RaCreate

The first three allow you to read and write ROIs to uncompressed, single file,
ICS v2 files. The last allows you to create an empty ICS file with given
datatype and dimensions - useful when you are processing large files and
need to create the output file or a scratch file.

Some existing functions needed to be modified in order to be able to open
the ICS file in the correct mode and obtain the FILE handle: contrary to
the existing library functions the ICS file is kept open until Ics_RaClose()
is called. Affected are:

- IcsOpenIcs
- IcsReadIcs
- IcsWriteIcs